refactor: moved the packages with a screen and viewmodel to a package called screens
