+++ app/src/main/java/uk/udemy/recordshop/ui/navigation/navgraphs/GenresGraph.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
-import uk.udemy.recordshop.ui.genres.GenresScreen
+import uk.udemy.recordshop.ui.screens.genres.GenresScreen
 import uk.udemy.recordshop.ui.navigation.Screens
 import uk.udemy.recordshop.ui.navigation.Tabs
