+++ app/src/main/java/uk/udemy/recordshop/ui/screens/addOrEditAlbum/AddOrEditAlbumViewModel.kt
@@ -1,4 +1,4 @@
-package uk.udemy.recordshop.ui.addOrEditAlbum
+package uk.udemy.recordshop.ui.screens.addOrEditAlbum
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
