+++ app/src/main/java/uk/udemy/recordshop/ui/screens/artist/artist/ArtistScreenContent.kt
@@ -1,4 +1,4 @@
-package uk.udemy.recordshop.ui.artist
+package uk.udemy.recordshop.ui.screens.artist.artist
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
